Update in Structure
