+++ src/Components/Certificate/Certificate.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Certificate = () => {
+    
+    return (
+        <a href="https://app.testdome.com/cert/6d00e3ccf7914b39b0c06ba39f6db8d1" className="testdome-certificate-stamp gold"><span className="testdome-certificate-name">    
+        Lingeshwaran Krishnan
+    
+</span><span className="testdome-certificate-test-name">PHP </span><span className="testdome-certificate-card-logo">TestDome<br />Certificate</span></a>
+    )
+}
+export default Certificate;
