--- src/Components/Portfolio.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Portfolio = () => {
-    return (
-            <section className="Portfolio_Section">
-                <div className="portfolio_1" >
-                        
-                </div>
-            </section>
-    )
-}
-export default Portfolio;
